Project/requirements.txt
########## requirements for cnnModel.py ###########
Python 3.9.12
Python 3.13.12


numpy==1.24.4
pandas==1.5.3
matplotlib==3.7.1
opencv-python==4.8.1.78
tensorflow==2.15.0
scikit-learn==1.3.2
Pillow==9.5.0
seaborn==0.12.2

###################################################


########## requirements for SVM.py ###############
Python 3.9.12
Python 3.13.12


numpy==1.24.4
matplotlib==3.7.1
numpy==1.24.4
opencv-python==4.8.1.78
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.3.2
scikit-image==0.21.0
joblib==1.3.2

##################################################


########## requirements for test_cnn & test_svm ###############
Python 3.9.12
Python 3.13.12


numpy==1.24.4
opencv-python==4.8.1.78
matplotlib==3.7.1
tensorflow==2.15.0
scikit-learn==1.3.2
seaborn==0.12.2
scikit-image==0.21.0
joblib==1.3.2

##############################################################